requirements.txt
octoparse
gspread
google-auth
google-auth-oauthlib
google-api-python-client
pandas
python-dotenv
